requirements.txt
aiogram
pymongo
pymongo[srv]
python-dotenv
markovify